Major performance improvement
Use the `beg end` provided by jit-lock-mode, and only update the related part.
We decrease `beg` by 1 to include the change caused by change at `beg`, which may
impact the spacing right before it. Similarly, we increase `end` by 1. However,
we should not delete overlays at `(1+ end)`, since that is the spacing between
`(1+ end)` and `(+ end 2)`, which is not changed by the current update. NOTE,
there is a known issue: if the buffer does not have a trailing eol, and you
delete everything after the last pangu spacing, that spacing will persist.