Fix problem in regexp
rx does not support "and", which says the text matches some regexp and some
other regexp simultaneously. Workaround was made to first try to match the
characters that should be excluded, then match chinese-two-byte. Thus the
position for chinese-two-byte will be nil if the character belongs both to
chinese-two-byte and excluded characters.